refactor: refactor tealium package
